added CRUD operations for tasks

+++ backend/routes/projects.js
@@ -1,13 +1,13 @@
-// Third-party packages.
+// Libraries.
 const express = require("express");
 const router = express.Router();
-// Mongoose models.
+// DB models.
 const { Project } = require("../models/index");
 const checkAuth = require("../middleware/check-auth");
 
 /* Get all projects. */
 router.get("/", checkAuth, (req, res) => {
-  Project.find()
+  Project.find({ creator: req.userData.userId })
     .then(projects => {
       const projectList = projects.map(project => {
         return {
@@ -19,15 +19,17 @@ router.get("/", checkAuth, (req, res) => {
     })
     .catch(err => {
       res.status(400).json({
+        message: "Unable to get projects.",
         error: err
       });
     });
 });
 
 /* Add a new project. */
-router.post("/", (req, res) => {
+router.post("/", checkAuth, (req, res) => {
   const project = new Project({
-    title: req.body.title
+    title: req.body.title,
+    creator: req.userData.userId
   });
   project
     .save()
@@ -39,37 +41,43 @@ router.post("/", (req, res) => {
     })
     .catch(err => {
       res.status(400).json({
+        message: "Unable to add this project.",
         error: err
       });
     });
 });
 
 /* Update an existing project. */
-router.put("/:id", (req, res) => {
-  Project.findOneAndUpdate({ _id: req.params.id }, { $set: req.body })
+router.put("/:id", checkAuth, (req, res) => {
+  Project.findOneAndUpdate(
+    { _id: req.params.id, creator: req.userData.userId },
+    { $set: req.body }
+  )
     .then(() => {
       res.status(200).json({
-        message: "Project updated."
+        message: "Project updated successfully."
       });
     })
     .catch(err => {
       res.status(400).json({
+        message: "Project updated failed.",
         error: err
       });
     });
 });
 
 /* Delete an existing project. */
-router.delete("/:id", (req, res) => {
-  Project.findOneAndDelete({ _id: req.params.id })
+router.delete("/:id", checkAuth, (req, res) => {
+  Project.findOneAndDelete({ _id: req.params.id, creator: req.userData.userId })
     .then(removedProject => {
       res.status(200).json({
-        message: "Project removed.",
+        message: "Project removed successfully.",
         removedProject: removedProject
       });
     })
     .catch(err => {
       res.status(400).json({
+        message: "Project removed failed.",
         error: err
       });
     });
